homePage dynamism
