+++ www/assets/front_office/homepage.ts
@@ -1,6 +1,6 @@
 require('./scss/homepage.scss');
 require('./ts/partials/layout.ts');
-import * as $ from 'jquery';
+require('./ts/components/slider');
 
 /*
 let slideIndex = 1;
@@ -31,42 +31,3 @@ function showSlides(n : number) {
     }
     slides[slideIndex-1].style.display = "flex";
 }*/
-
-$.fn.nextOrFirst = function(selector){
-    var next = this.next(selector);
-    return (next.length) ? next : this.prevAll(selector).last();
-};
-
-$.fn.prevOrLast = function(selector){
-    var prev = this.prev(selector);
-    return (prev.length) ? prev : this.nextAll(selector).last();
-};
-
-function slider() {
-    var activeSlide = $(".active");
-
-    activeSlide
-        .removeClass("active")
-        .nextOrFirst()
-        .addClass("active");
-}
-setInterval(slider, 6000);
-
-function controls() {
-    var control = $(".controls");
-
-    control.on('click', '.prev', function() {
-        $(".active")
-            .removeClass("active")
-            .prevOrLast()
-            .addClass("active");
-    });
-
-    control.on('click', '.next', function() {
-        $(".active")
-            .removeClass("active")
-            .nextOrFirst()
-            .addClass("active");
-    });
-}
-controls();
