+++ www/assets/front_office/scss/components/slider.scss
@@ -1,0 +1,87 @@
+.slideshow-container {
+  position: relative;
+  margin: auto;
+  height: 100vh;
+
+  article{
+
+    p{
+      margin-bottom: 30px;
+    }
+  }
+
+  .prev, .next {
+    cursor: pointer;
+    position: absolute;
+    top: 50%;
+    transform: translateY(-50%);
+    width: auto;
+    padding: 16px;
+    color: #2aad38;
+    font-weight: bold;
+    font-size: 2rem;
+    transition: 0.5s ease;
+    border-radius: 0 3px 3px 0;
+    user-select: none;
+    z-index: 11;
+  }
+
+  .next {
+    right: 0;
+    border-radius: 3px 0 0 3px;
+  }
+
+  .prev:hover, .next:hover {
+    background-color: rgba(0,0,0,0.8);
+  }
+
+}
+
+/*
+.fade {
+  -webkit-animation-name: fade;
+  -webkit-animation-duration: 1.5s;
+  animation-name: fade;
+  animation-duration: 1.5s;
+}
+
+@-webkit-keyframes fade {
+  from {opacity: .4}
+  to {opacity: 1}
+}
+
+@keyframes fade {
+  from {opacity: .4}
+  to {opacity: 1}
+}
+*/
+
+.slide {
+  position: absolute;
+  opacity: 0;
+  z-index: 9;
+
+  img {
+    overflow: hidden;
+    object-fit: cover;
+    height: auto;
+  }
+
+  span {
+    background: rgba(0,0,0,0.5);
+    bottom: 10px; left: 10px;
+    display: block;
+    padding: 15px 0;
+    position: absolute;
+    overflow: hidden;
+    text-align: center;
+    width: 600px;
+  }
+}
+
+.active {
+  display: block;
+  opacity: 1;
+  transition: opacity 1s ease;
+  z-index: 10;
+}
