+++ www/assets/front_office/scss/homepage.scss
@@ -1,59 +1,10 @@
 @import 'app.scss';
+@import 'components/slider';
 
 main{
-  .slideshow-container {
-    position: relative;
-    margin: auto;
 
-    article{
 
-      p{
-        margin-bottom: 30px;
-      }
-    }
-  }
-
-  .prev, .next {
-    cursor: pointer;
-    position: absolute;
-    top: 50%;
-    width: auto;
-    padding: 16px;
-    color: #2aad38;
-    font-weight: bold;
-    font-size: 2rem;
-    transition: 0.5s ease;
-    border-radius: 0 3px 3px 0;
-    user-select: none;
-  }
-
-  .next {
-    right: 0;
-    border-radius: 3px 0 0 3px;
-  }
-
-  .prev:hover, .next:hover {
-    background-color: rgba(0,0,0,0.8);
-  }
-
-  .fade {
-    -webkit-animation-name: fade;
-    -webkit-animation-duration: 1.5s;
-    animation-name: fade;
-    animation-duration: 1.5s;
-  }
-
-  @-webkit-keyframes fade {
-    from {opacity: .4}
-    to {opacity: 1}
-  }
-
-  @keyframes fade {
-    from {opacity: .4}
-    to {opacity: 1}
-  }
-
-  #entreprise, #zerodechet{
+  .homePage {
     background: $lightGrey;
     article{
       padding: 0 0 0 $basicPadding;
